refactor: 규칙 재구성에 따른 패키지 이동
