+++ src/main/kotlin/omok/domain/rule/place/InvalidPositionRule.kt
@@ -1,9 +1,9 @@
-package omok.domain.rule
+package omok.domain.rule.place
 
 import omok.domain.omokboard.OmokBoard
 import omok.domain.player.PlayerStone
 
-class InvalidPositionRule : OmokRule {
+class InvalidPositionRule : PlaceRule {
     override fun canPlace(
         omokBoard: OmokBoard,
         playerStone: PlayerStone,
