+++ src/main/kotlin/omok/domain/rule/place/PlaceRule.kt
@@ -1,9 +1,9 @@
-package omok.domain.rule
+package omok.domain.rule.place
 
 import omok.domain.omokboard.OmokBoard
 import omok.domain.player.PlayerStone
 
-interface OmokRule {
+interface PlaceRule {
     fun canPlace(
         omokBoard: OmokBoard,
         playerStone: PlayerStone,
